improve adding students to classroom
I had to use on over classroom/{uid} instead of once in order to update the list of students in real time (while adding and removing students). Since we use on I have to remove the listener to this specific child. Thus, there is a new action to remove child listener